better error handling for null models
This is an interesting problem! If you make an empty list, all objects in the list are null. However if you assign one of the elements to be NULL, that deletes it from the list. Every time there was a try error it was assigning a null value to the mouts list object. So every time the model fails it shortens the list object, hence why the out of bounds error occurs.

Here is a fixed version, I just use if statements to replace the NULL